// packages/theme/src/sys-tokens/colors/jooc-sys-color-tokens.types.ts
import { type JoocRefPaletteTokensHexCodes } from "../../ref-tokens/palette";

export type JoocSysColorTokensType = {
  background: JoocRefPaletteTokensHexCodes;
  onBackground: JoocRefPaletteTokensHexCodes;
  surface: JoocRefPaletteTokensHexCodes;
  onSurface: JoocRefPaletteTokensHexCodes;
  surfaceVariant: JoocRefPaletteTokensHexCodes;
  onSurfaceVariant: JoocRefPaletteTokensHexCodes;
  surfacePrimary: JoocRefPaletteTokensHexCodes;
  onSurfacePrimary: JoocRefPaletteTokensHexCodes;
  surfacePrimaryVariant: JoocRefPaletteTokensHexCodes;
  onSurfacePrimaryVariant: JoocRefPaletteTokensHexCodes;
  surfaceError: JoocRefPaletteTokensHexCodes;
  onSurfaceError: JoocRefPaletteTokensHexCodes;
  surfaceSuccess: JoocRefPaletteTokensHexCodes;
  onSurfaceSuccess: JoocRefPaletteTokensHexCodes;
  primary: JoocRefPaletteTokensHexCodes;
  primaryHover: JoocRefPaletteTokensHexCodes;
  primaryPress: JoocRefPaletteTokensHexCodes;
  onPrimary: JoocRefPaletteTokensHexCodes;
  secondary: JoocRefPaletteTokensHexCodes;
  secondaryHover: JoocRefPaletteTokensHexCodes;
  secondaryPress: JoocRefPaletteTokensHexCodes;
  onSecondary: JoocRefPaletteTokensHexCodes;
  terciary: JoocRefPaletteTokensHexCodes;
  terciaryHover: JoocRefPaletteTokensHexCodes;
  terciaryPress: JoocRefPaletteTokensHexCodes;
  onTerciary: JoocRefPaletteTokensHexCodes;
  critical: JoocRefPaletteTokensHexCodes;
  criticalHover: JoocRefPaletteTokensHexCodes;
  criticalPress: JoocRefPaletteTokensHexCodes;
  onCritical: JoocRefPaletteTokensHexCodes;
  alert: JoocRefPaletteTokensHexCodes;
  onAlert: JoocRefPaletteTokensHexCodes;
  success: JoocRefPaletteTokensHexCodes;
  onSuccess: JoocRefPaletteTokensHexCodes;
  disabled: JoocRefPaletteTokensHexCodes;
  onDisabled: JoocRefPaletteTokensHexCodes;
};